픽업 홈에 연결
